pyproject.toml
[tool.poetry]
name = "kafka-python-app"
version = "0.1.0"
description = "A simple Kafka producer/consumer application"
authors = ["Your Name <your.email@example.com>"]

[[tool.poetry.packages]]
include = "app"

[tool.poetry.dependencies]
python = "^3.13"
kafka-python = "^2.0.2"
uvicorn = "^0.35.0"
pydantic = "^2.11.7"
fastapi = "^0.116.1"
requests = "^2.32.4"
confluent-kafka = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
